# openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseConsultingLeadResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ConsultingLeadResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListAuditView" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/AuditView"
            },
            "type" : "array"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListUserSummary" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/UserSummary"
            },
            "type" : "array"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseProfile" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/Profile"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseVoid" : {
        "properties" : {
          "data" : {
            "type" : "object"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "AuditView" : {
        "properties" : {
          "actor" : {
            "type" : "string"
          },
          "httpMethod" : {
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "requestPath" : {
            "type" : "string"
          },
          "statusCode" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ChangePassword" : {
        "properties" : {
          "currentPassword" : {
            "minLength" : 1,
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 64,
            "minLength" : 8,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ChangeRoleRequest" : {
        "properties" : {
          "role" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ConsultingLeadResponse" : {
        "properties" : {
          "diagnosisId" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "maskedEmail" : {
            "type" : "string"
          },
          "maskedPhone" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "fieldErrors" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FieldError" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "rejectedValue" : {
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "Logout" : {
        "properties" : {
          "refreshToken" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Profile" : {
        "properties" : {
          "createdAt" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "emailVerified" : {
            "type" : "boolean"
          },
          "nickname" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RequestConsultingRequest" : {
        "properties" : {
          "consentVersion" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "contactName" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactPhone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "diagnosisId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          },
          "privacyConsent" : {
            "type" : "boolean"
          },
          "sensitiveInfoConsent" : {
            "type" : "boolean"
          },
          "serviceLimitAcknowledged" : {
            "type" : "boolean"
          }
        },
        "required" : [ "diagnosisId" ],
        "type" : "object"
      },
...
      "UpdateNickname" : {
        "properties" : {
          "nickname" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserSummary" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "emailVerified" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "nickname" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "로그인(POST /api/v1/auth/login) 응답의 accessToken",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "1인 제조기업·소공인 제품 인증 준비도 자가진단 및 컨설팅 연계 서비스.\n\n• 모든 응답은 ApiResponse 봉투(success/data/error/traceId/timestamp)로 감쌉니다.\n• 준비도 점수는 인증 합격·불합격 판정이 아니라 준비 상태를 확인하는 사전 점검 지표입니다.\n• 인증이 필요한 엔드포인트는 우상단 Authorize에 로그인 accessToken을 넣고 호출하세요.\n",
    "license" : {
      "name" : "팀 육성사이다"
    },
    "title" : "OpenCertFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/audit-logs" : {
      "get" : {
        "operationId" : "list_11",
        "parameters" : [ {
          "in" : "query",
          "name" : "actor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAuditView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-audit-log-controller" ]
      }
    },
...
    "/api/v1/admin/users" : {
      "get" : {
        "operationId" : "list_7",
        "parameters" : [ {
          "in" : "query",
          "name" : "role",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListUserSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
    "/api/v1/admin/users/{userId}/role" : {
      "patch" : {
        "operationId" : "changeRole",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeRoleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseProfile"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
...
    "/api/v1/consulting-leads" : {
      "post" : {
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RequestConsultingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseConsultingLeadResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "상담 신청(공개)" ]
      }
    },
...
    "/api/v1/me/nickname" : {
      "patch" : {
        "operationId" : "updateNickname",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateNickname"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseProfile"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "마이페이지" ]
      }
    },
...
    "/api/v1/me/password" : {
      "patch" : {
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePassword"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "마이페이지" ]
      }
    },
    "/api/v1/me/session" : {
      "delete" : {
        "operationId" : "logout",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Logout"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "마이페이지" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : ""
  } ],
...
}